+++ @angular-generic-table/column-settings/index.ts
@@ -1,0 +1,2 @@
+
+export { ColumnSettingsModule } from './column-settings.module';