+++ @angular-generic-table/column-settings/interfaces/gt-column-settings-texts.ts
@@ -1,0 +1,4 @@
+export interface GtColumnSettingsTexts {
+  title:string;
+  help:string;
+}