+++ @angular-generic-table/core/pipes/gt-visible.pipe.ts
@@ -10,7 +10,7 @@ export class GtVisiblePipe<R extends GtRow> implements PipeTransform {
 
   // TODO: move to helper functions
   /** Sort by column order */
-  private getColumnOrder = function(a,b) {
+  private getColumnOrder = function(a:any,b:any) {
     if (a.columnOrder < b.columnOrder)
       return -1;
     if (a.columnOrder > b.columnOrder || typeof a.columnOrder === 'undefined')