+++ src/app/app.module.ts
@@ -15,8 +15,6 @@ import { ExemplifyModule } from "angular-exemplify";
 import { LocalizationComponent} from './localization/localization.component';
 import { CustomColumnComponent, NameComponent, AgeComponent } from './custom-column/custom-column.component';
 
-import {DragulaModule} from 'ng2-dragula'
-
 /** Only needed when using ng2-translate */
 import { TranslateModule, TranslateLoader} from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
@@ -27,7 +25,7 @@ export function createTranslateLoader(http: Http) {
 
 /** Import generic table module */
 import { GenericTableModule } from '../../@angular-generic-table/core';
-import {GtColumnPipe, GtColumnSettingsComponent} from '../../@angular-generic-table/column-settings/components/gt-column-settings/gt-column-settings.component';
+import { ColumnSettingsModule } from "../../@angular-generic-table/column-settings/column-settings.module";
 
 @NgModule({
   declarations: [
@@ -40,15 +38,14 @@ import {GtColumnPipe, GtColumnSettingsComponent} from '../../@angular-generic-ta
     AgeComponent,
     BasicComponent,
     ExamplesComponent,
-    LocalizationComponent,
-    GtColumnSettingsComponent,
-    GtColumnPipe
+    LocalizationComponent
   ],
   imports: [
     BrowserModule,
     FormsModule,
     HttpModule,
     GenericTableModule, /** ADD THIS LINE TO YOUR APP MODULE! */
+    ColumnSettingsModule,
     AppRoutingModule, /** holds routes used in examples */
     ExemplifyModule, /** used for generating examples */
     /** translate module only needed for localization when using ngx */
@@ -58,8 +55,7 @@ import {GtColumnPipe, GtColumnSettingsComponent} from '../../@angular-generic-ta
         useFactory: (createTranslateLoader),
         deps: [Http]
       }
-    }),
-    DragulaModule
+    })
   ],
   /** add components used by your table i.e. for expanding rows etc. as entry components */
   entryComponents: [